dom/tests/mochitest/pointerlock/test_pointerlockerror.html

<!DOCTYPE HTML>
<html>
<!--https://bugzilla.mozilla.org/show_bug.cgi?id=633602-->
<head>
  <title>Bug 633602</title>
  <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js">
  </script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js">
  </script>
  <script type="application/javascript" src="testPrefs.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body onload="start();">
  <a target="_blank"
    href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">
    Mozilla Bug 633602
  </a>

  <pre id="test">
    <script type="text/javascript">
      /*
       * Test for Bug 633602
       * Make sure pointerlockerror event fires.
       */

      SimpleTest.waitForExplicitFinish();

      document.addEventListener("mozpointerlockerror", function (e) {
        ok(true, "pointerlockerror event should fire.");
        testPrefs.restore();
        SimpleTest.finish();
      }, false);

      function start() {
        SimpleTest.waitForFocus(function() {
          // element not in the DOM, not fullscreen, should fail to lock
          testPrefs.init();
          div = document.createElement("div");
          div.mozRequestPointerLock();
        });
      }
    </script>
  </pre>
</body>
</html>
